moved delete widget to saperate component
